refactor(ui): dedupe speech-text frontend and remove dead code
