--- ui/src/components/ui/ProgressRing.tsx
@@ -1,73 +1,0 @@
-import { useId } from "react";
-import { motion } from "framer-motion";
-
-interface ProgressRingProps {
-  progress: number;
-  size?: number;
-  strokeWidth?: number;
-  className?: string;
-}
-
-export function ProgressRing({
-  progress,
-  size = 48,
-  strokeWidth = 4,
-  className = "",
-}: ProgressRingProps) {
-  const gradientId = useId();
-  const radius = (size - strokeWidth) / 2;
-  const circumference = radius * 2 * Math.PI;
-  const offset = circumference - (progress / 100) * circumference;
-
-  return (
-    <div
-      className={`relative ${className}`}
-      style={{ width: size, height: size }}
-    >
-      <svg width={size} height={size} className="transform -rotate-90">
-        {/* Background circle */}
-        <circle
-          cx={size / 2}
-          cy={size / 2}
-          r={radius}
-          fill="none"
-          stroke="var(--border-muted)"
-          strokeWidth={strokeWidth}
-        />
-        {/* Progress circle */}
-        <motion.circle
-          cx={size / 2}
-          cy={size / 2}
-          r={radius}
-          fill="none"
-          stroke={`url(#${gradientId})`}
-          strokeWidth={strokeWidth}
-          strokeLinecap="round"
-          initial={{ strokeDashoffset: circumference }}
-          animate={{ strokeDashoffset: offset }}
-          transition={{ duration: 0.5, ease: "easeOut" }}
-          style={{
-            strokeDasharray: circumference,
-          }}
-        />
-        <defs>
-          <linearGradient
-            id={gradientId}
-            x1="0%"
-            y1="0%"
-            x2="100%"
-            y2="0%"
-          >
-            <stop offset="0%" stopColor="var(--text-subtle)" />
-            <stop offset="100%" stopColor="var(--text-secondary)" />
-          </linearGradient>
-        </defs>
-      </svg>
-      <div className="absolute inset-0 flex items-center justify-center">
-        <span className="text-xs font-semibold text-[var(--text-secondary)]">
-          {Math.round(progress)}%
-        </span>
-      </div>
-    </div>
-  );
-}
