+++ ui/src/features/chat/route.tsx
@@ -2,6 +2,7 @@ import type { ModelInfo } from "@/api";
 import { PageShell } from "@/components/PageShell";
 import { VIEW_CONFIGS } from "@/types";
 import { ChatPlayground } from "@/features/chat/components/ChatPlayground";
+import type { ModelDownloadProgressMap } from "@/features/models/downloadProgress";
 import {
   CHAT_PREFERRED_MODELS,
   getChatRouteModelLabel,
@@ -15,16 +16,7 @@ interface ChatPageProps {
   models: ModelInfo[];
   selectedModel: string | null;
   loading: boolean;
-  downloadProgress: Record<
-    string,
-    {
-      percent: number;
-      currentFile: string;
-      status: string;
-      downloadedBytes: number;
-      totalBytes: number;
-    }
-  >;
+  downloadProgress: ModelDownloadProgressMap;
   onDownload: (variant: string) => void;
   onCancelDownload?: (variant: string) => void;
   onLoad: (variant: string) => void;
