+++ ui/src/features/models/downloadProgress.ts
@@ -1,0 +1,9 @@
+export interface ModelDownloadProgressEntry {
+  percent: number;
+  currentFile: string;
+  status: string;
+  downloadedBytes: number;
+  totalBytes: number;
+}
+
+export type ModelDownloadProgressMap = Record<string, ModelDownloadProgressEntry>;
