+++ ui/src/features/voice-cloning/route.tsx
@@ -3,6 +3,7 @@ import { useNavigate } from "react-router-dom";
 import type { ModelInfo } from "@/api";
 import { PageHeader, PageShell } from "@/components/PageShell";
 import { VIEW_CONFIGS } from "@/types";
+import type { ModelDownloadProgressMap } from "@/features/models/downloadProgress";
 import {
   VOICE_CLONING_PREFERRED_MODELS,
   resolvePreferredRouteModel,
@@ -15,16 +16,7 @@ interface VoiceCloningPageProps {
   models: ModelInfo[];
   selectedModel: string | null;
   loading: boolean;
-  downloadProgress: Record<
-    string,
-    {
-      percent: number;
-      currentFile: string;
-      status: string;
-      downloadedBytes: number;
-      totalBytes: number;
-    }
-  >;
+  downloadProgress: ModelDownloadProgressMap;
   onDownload: (variant: string) => void;
   onCancelDownload?: (variant: string) => void;
   onLoad: (variant: string) => void;
